[DI] Allow explicit taskLauncher for torchscript interpreter
Pull Request resolved: https://github.com/pytorch/pytorch/pull/46444

By default, TorchScript execution is single threaded and uses the caller's thread pool. For the use case of distributed inference, we hope there is a way to customize the behavior where the  interpreter in torch script can be executed in other places. This diff allows an explicit taskLauncher for torchscript interpreter.

ghstack-source-id: 114546287

Differential Revision: [D24114874](https://our.internmc.facebook.com/intern/diff/D24114874/)